backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pandas==2.2.1
numpy==1.26.4
httpx==0.26.0
pydantic==1.10.14
python-dotenv==1.0.1
SQLAlchemy==2.0.29
psycopg2-binary==2.9.9
